+++ examples/common-ggml.h
@@ -16,7 +16,3 @@ bool ggml_common_quantize_0(
         const ggml_ftype ftype,
         const std::vector<std::string> & to_quant,
         const std::vector<std::string> & to_skip);
-
-// these will move to ggml when ready
-void        ggml_cgraph_export(const struct ggml_cgraph * cgraph, const char * fname);
-ggml_cgraph ggml_cgraph_import(const char * fname, struct ggml_context ** ctx_data, struct ggml_context ** ctx_eval);
